# backend/contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateNotificationDto": {
        "properties": {},
        "type": "object"
      },
...
      "TestEmailDto": {
        "properties": {},
        "type": "object"
      },
...
      "UpdateEmailConfigDto": {
        "properties": {},
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "Mock backend API for Gush.",
    "title": "Gush Backend",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/email": {
      "get": {
        "operationId": "AdminEmailController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "AdminEmailController_save",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailConfigDto"
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/admin/email/test": {
      "post": {
        "operationId": "AdminEmailController_test",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/admin/notifications": {
      "get": {
        "operationId": "AdminNotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "AdminNotificationsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/admin/notifications/{id}": {
      "delete": {
        "operationId": "AdminNotificationsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "SearchController_search",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "genre",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "adult",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/search/hot": {
      "get": {
        "operationId": "SearchController_hot",
        "parameters": [
          {
            "in": "query",
            "name": "adult",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/search/keywords": {
      "get": {
        "operationId": "SearchController_keywords",
        "parameters": [
          {
            "in": "query",
            "name": "adult",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/search/suggest": {
      "get": {
        "operationId": "SearchController_suggest",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "adult",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "servers": [],
...
}