# backend/src/modules/admin/admin-system/controllers/admin-notifications.controller.ts
import {
  BadRequestException,
  Body,
  Controller,
  Delete,
  Get,
  NotFoundException,
  Param,
  Post,
  Req,
  UseGuards,
} from "@nestjs/common";
import { ApiBody } from "@nestjs/swagger";
import { Request } from "express";
import { PrismaService } from "../../../../common/prisma/prisma.service";
import {
  buildPaginationResult,
  calculateOffset,
  parsePaginationParams,
} from "../../../../common/utils/pagination";
import { AdminAuthGuard } from "../../guards/admin-auth.guard";
import {
  CreateNotificationDto,
  NotificationPayloadInput,
} from "../dtos/admin-system.dto";

type NotificationRequestBody = CreateNotificationDto & NotificationPayloadInput;

function extractNotificationPayload(body: NotificationRequestBody): NotificationPayloadInput {
  return body?.notification || body;
}

@Controller("admin/notifications")
@UseGuards(AdminAuthGuard)
export class AdminNotificationsController {
  constructor(private readonly prisma: PrismaService) {}

  @Get()
  async list(@Req() req: Request) {
    const { page, pageSize } = parsePaginationParams(req.query);
    const offset = calculateOffset(page, pageSize);

    const [notifications, total] = await Promise.all([
      this.prisma.notification.findMany({
        orderBy: { createdAt: "desc" },
        take: pageSize,
        skip: offset,
      }),
      this.prisma.notification.count(),
    ]);

    return buildPaginationResult(notifications, total, page, pageSize);
  }

  @Post()
  @ApiBody({ type: CreateNotificationDto, required: false })
  async create(@Body() body: NotificationRequestBody) {
    const payload = extractNotificationPayload(body);
    if (!payload.title) {
      throw new BadRequestException("Missing title");
    }

    if (payload.broadcast) {
      const batchSize = 1000;
      let skip = 0;
      let totalCreated = 0;

      while (true) {
        const users = await this.prisma.user.findMany({
          select: { id: true },
          take: batchSize,
          skip,
        });

        if (users.length === 0) {
          break;
        }

        await this.prisma.notification.createMany({
          data: users.map((user) => ({
            id: `N_${user.id}_${Date.now()}`,
            userId: user.id,
            type: payload.type || "PROMO",
            title: payload.title || "",
            message: payload.message || "",
            seriesId: payload.seriesId || null,
            episodeId: payload.episodeId || null,
            read: false,
            createdAt: new Date(),
          })),
        });

        totalCreated += users.length;
        skip += batchSize;
      }

      return { ok: true, count: totalCreated };
    }

    const userId = payload.userId;
    if (!userId) {
      throw new BadRequestException("Missing userId");
    }

    const notification = await this.prisma.notification.create({
      data: {
        id: `N_${userId}_${Date.now()}`,
        userId,
        type: payload.type || "PROMO",
        title: payload.title,
        message: payload.message || "",
        seriesId: payload.seriesId || null,
        episodeId: payload.episodeId || null,
        read: false,
        createdAt: new Date(),
      },
    });

    return { notification };
  }

  @Delete(":id")
  async remove(@Param("id") id: string) {
    if (!id) {
      throw new BadRequestException("Missing notification id");
    }

    const existing = await this.prisma.notification.findUnique({ where: { id } });
    if (!existing) {
      throw new NotFoundException("Notification not found");
    }

    await this.prisma.notification.delete({ where: { id } });
    return { ok: true };
  }
}